[Flare] Ensure updateEventListeners updates in commit phase
