+++ packages/react-reconciler/src/ReactFiberCompleteWork.js
@@ -9,12 +9,7 @@
 
 import type {Fiber} from './ReactFiber';
 import type {ExpirationTime} from './ReactFiberExpirationTime';
-import type {
-  ReactEventResponder,
-  ReactEventResponderInstance,
-  ReactFundamentalComponentInstance,
-  ReactEventResponderListener,
-} from 'shared/ReactTypes';
+import type {ReactFundamentalComponentInstance} from 'shared/ReactTypes';
 import type {FiberRoot} from './ReactFiberRoot';
 import type {
   Instance,
@@ -31,7 +26,6 @@ import type {SuspenseContext} from './ReactFiberSuspenseContext';
 
 import {now} from './SchedulerWithReactIntegration';
 
-import {REACT_RESPONDER_TYPE} from 'shared/ReactSymbols';
 import {
   IndeterminateComponent,
   FunctionComponent,
@@ -79,8 +73,6 @@ import {
   createContainerChildSet,
   appendChildToContainerChildSet,
   finalizeContainerChildren,
-  mountResponderInstance,
-  unmountResponderInstance,
   getFundamentalComponentInstance,
   mountFundamentalComponent,
   cloneFundamentalInstance,
@@ -92,8 +84,6 @@ import {
   getHostContext,
   popHostContainer,
 } from './ReactFiberHostContext';
-import {NoWork} from './ReactFiberExpirationTime';
-import {createResponderInstance} from './ReactFiberEvents';
 import {
   suspenseStackCursor,
   InvisibleParentSuspenseContext,
@@ -133,10 +123,7 @@ import {
 import {createFundamentalStateInstance} from './ReactFiberFundamental';
 import {Never} from './ReactFiberExpirationTime';
 import {resetChildFibers} from './ReactChildFiber';
-import warning from 'shared/warning';
-
-const emptyObject = {};
-const isArray = Array.isArray;
+import {updateEventListeners} from './ReactFiberEvents';
 
 function markUpdate(workInProgress: Fiber) {
   // Tag the fiber with an update effect. This turns a Placement into
@@ -693,14 +680,8 @@ function completeWork(
         if (enableFlareAPI) {
           const prevListeners = current.memoizedProps.listeners;
           const nextListeners = newProps.listeners;
-          const instance = workInProgress.stateNode;
           if (prevListeners !== nextListeners) {
-            updateEventListeners(
-              nextListeners,
-              instance,
-              rootContainerInstance,
-              workInProgress,
-            );
+            markUpdate(workInProgress);
           }
         }
 
@@ -742,12 +723,7 @@ function completeWork(
             const instance = workInProgress.stateNode;
             const listeners = newProps.listeners;
             if (listeners != null) {
-              updateEventListeners(
-                listeners,
-                instance,
-                rootContainerInstance,
-                workInProgress,
-              );
+              updateEventListeners(listeners, instance, workInProgress);
             }
           }
         } else {
@@ -764,12 +740,7 @@ function completeWork(
           if (enableFlareAPI) {
             const listeners = newProps.listeners;
             if (listeners != null) {
-              updateEventListeners(
-                listeners,
-                instance,
-                rootContainerInstance,
-                workInProgress,
-              );
+              updateEventListeners(listeners, instance, workInProgress);
             }
           }
 
@@ -1256,158 +1227,6 @@ function completeWork(
   }
 
   return null;
-}
-
-function mountEventResponder(
-  responder: ReactEventResponder<any, any>,
-  responderProps: Object,
-  instance: Instance,
-  rootContainerInstance: Container,
-  fiber: Fiber,
-  respondersMap: Map<
-    ReactEventResponder<any, any>,
-    ReactEventResponderInstance<any, any>,
-  >,
-) {
-  let responderState = emptyObject;
-  const getInitialState = responder.getInitialState;
-  if (getInitialState !== null) {
-    responderState = getInitialState(responderProps);
-  }
-  const responderInstance = createResponderInstance(
-    responder,
-    responderProps,
-    responderState,
-    instance,
-    fiber,
-  );
-  mountResponderInstance(
-    responder,
-    responderInstance,
-    responderProps,
-    responderState,
-    instance,
-    rootContainerInstance,
-  );
-  respondersMap.set(responder, responderInstance);
-}
-
-function updateEventListener(
-  listener: ReactEventResponderListener<any, any>,
-  fiber: Fiber,
-  visistedResponders: Set<ReactEventResponder<any, any>>,
-  respondersMap: Map<
-    ReactEventResponder<any, any>,
-    ReactEventResponderInstance<any, any>,
-  >,
-  instance: Instance,
-  rootContainerInstance: Container,
-): void {
-  let responder;
-  let props;
-
-  if (listener) {
-    responder = listener.responder;
-    props = listener.props;
-  }
-  invariant(
-    responder && responder.$$typeof === REACT_RESPONDER_TYPE,
-    'An invalid value was used as an event listener. Expect one or many event ' +
-      'listeners created via React.unstable_useResponder().',
-  );
-  const listenerProps = ((props: any): Object);
-  if (visistedResponders.has(responder)) {
-    // show warning
-    if (__DEV__) {
-      warning(
-        false,
-        'Duplicate event responder "%s" found in event listeners. ' +
-          'Event listeners passed to elements cannot use the same event responder more than once.',
-        responder.displayName,
-      );
-    }
-    return;
-  }
-  visistedResponders.add(responder);
-  const responderInstance = respondersMap.get(responder);
-
-  if (responderInstance === undefined) {
-    // Mount
-    mountEventResponder(
-      responder,
-      listenerProps,
-      instance,
-      rootContainerInstance,
-      fiber,
-      respondersMap,
-    );
-  } else {
-    // Update
-    responderInstance.props = listenerProps;
-    responderInstance.fiber = fiber;
-  }
-}
-
-function updateEventListeners(
-  listeners: any,
-  instance: Instance,
-  rootContainerInstance: Container,
-  fiber: Fiber,
-): void {
-  const visistedResponders = new Set();
-  let dependencies = fiber.dependencies;
-  if (listeners != null) {
-    if (dependencies === null) {
-      dependencies = fiber.dependencies = {
-        expirationTime: NoWork,
-        firstContext: null,
-        responders: new Map(),
-      };
-    }
-    let respondersMap = dependencies.responders;
-    if (respondersMap === null) {
-      respondersMap = new Map();
-    }
-    if (isArray(listeners)) {
-      for (let i = 0, length = listeners.length; i < length; i++) {
-        const listener = listeners[i];
-        updateEventListener(
-          listener,
-          fiber,
-          visistedResponders,
-          respondersMap,
-          instance,
-          rootContainerInstance,
-        );
-      }
-    } else {
-      updateEventListener(
-        listeners,
-        fiber,
-        visistedResponders,
-        respondersMap,
-        instance,
-        rootContainerInstance,
-      );
-    }
-  }
-  if (dependencies !== null) {
-    const respondersMap = dependencies.responders;
-    if (respondersMap !== null) {
-      // Unmount
-      const mountedResponders = Array.from(respondersMap.keys());
-      for (let i = 0, length = mountedResponders.length; i < length; i++) {
-        const mountedResponder = mountedResponders[i];
-        if (!visistedResponders.has(mountedResponder)) {
-          const responderInstance = ((respondersMap.get(
-            mountedResponder,
-          ): any): ReactEventResponderInstance<any, any>);
-          unmountResponderInstance(responderInstance);
-          respondersMap.delete(mountedResponder);
-        }
-      }
-    }
-  }
 }
 
 export {completeWork};
